+++ jeap-process-archive-type-registry-maven-plugin/src/main/java/ch/admin/bit/jeap/processarchive/registry/ArchiveTypeRegistryVerifierMojo.java
@@ -1,6 +1,8 @@
 package ch.admin.bit.jeap.processarchive.registry;
 
 import ch.admin.bit.jeap.processarchive.avro.plugin.compiler.ImportClassLoader;
+import ch.admin.bit.jeap.processarchive.registry.git.GitClient;
+import ch.admin.bit.jeap.processarchive.registry.git.GitClientException;
 import ch.admin.bit.jeap.processarchive.registry.verifier.DescriptorDirectoryValidator;
 import ch.admin.bit.jeap.processarchive.registry.verifier.ValidationContext;
 import ch.admin.bit.jeap.processarchive.registry.verifier.common.ValidationResult;
@@ -15,8 +17,6 @@ import org.apache.maven.plugins.annotations.Mojo;
 import org.apache.maven.plugins.annotations.Parameter;
 import org.apache.maven.plugins.annotations.ResolutionScope;
 import org.apache.maven.project.MavenProject;
-import org.eclipse.jgit.api.Git;
-import org.eclipse.jgit.api.errors.GitAPIException;
 
 import java.io.File;
 import java.io.IOException;
@@ -31,6 +31,9 @@ public class ArchiveTypeRegistryVerifierMojo extends AbstractMojo {
     @SuppressWarnings("unused")
     @Parameter(name = "gitUrl")
     private String gitUrl;
+    @SuppressWarnings("unused")
+    @Parameter(name= "archiveTypeRepoGitTokenEnvVariableName", defaultValue = "ARCHIVE_TYPE_REPO_GIT_TOKEN")
+    private String archiveTypeRepoGitTokenEnvVariableName;
     @SuppressWarnings("unused")
     @Getter(AccessLevel.PROTECTED)
     @Parameter(defaultValue = "${project}", required = true, readonly = true)
@@ -64,14 +67,12 @@ public class ArchiveTypeRegistryVerifierMojo extends AbstractMojo {
         try {
             File tempDir = Files.createTempDirectory("master").toFile();
             FileUtils.forceDeleteOnExit(tempDir);
-            Git.cloneRepository()
-                    .setURI(gitUrl)
-                    .setBranch("master")
-                    .setDirectory(tempDir)
-                    .call();
+            GitClient gitClient = new GitClient(gitUrl, archiveTypeRepoGitTokenEnvVariableName, getLog());
+            gitClient.cloneAtBranch("master", tempDir);
             return new File(tempDir, descriptorDirectory.getName());
-        } catch (IOException | GitAPIException e) {
+        } catch (IOException | GitClientException e) {
             throw new MojoExecutionException("Cannot checkout old repo", e);
         }
     }
+
 }