+++ jeap-process-archive-type-registry-maven-plugin/src/main/java/ch/admin/bit/jeap/processarchive/registry/git/GitClientException.java
@@ -1,0 +1,13 @@
+package ch.admin.bit.jeap.processarchive.registry.git;
+
+public class GitClientException extends RuntimeException{
+
+    public GitClientException(String message) {
+        super(message);
+    }
+
+    public GitClientException(String message, Throwable cause) {
+        super(message, cause);
+    }
+
+}