+++ jeap-process-archive-type-registry-maven-plugin/src/main/java/ch/admin/bit/jeap/processarchive/registry/git/ProcessBuilderFactory.java
@@ -1,0 +1,13 @@
+package ch.admin.bit.jeap.processarchive.registry.git;
+
+/**
+ * Factory class to create instances of {@link ProcessBuilder}.
+ * This allows for easier testing and mocking of process execution.
+ */
+public class ProcessBuilderFactory {
+
+    ProcessBuilder createProcessBuilder(String... command) {
+        return new ProcessBuilder(command);
+    }
+
+}